requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
redis==5.2.0
openai==1.59.9
httpx>=0.28.0
python-dotenv==1.0.1
